+++ Extension/src/pages/options/components/Filters/AddCustomModal/AddCustomModal.jsx
@@ -128,6 +128,9 @@ const AddCustomModal = observer(({
             } else if (result.errorAlreadyExists) {
                 setError(translator.getMessage('options_antibanner_custom_filter_already_exists'));
                 setStepToRender(STEPS.ERROR);
+            } else if (result.errorHtmlContent) {
+                setError(translator.getMessage('options_add_custom_filter_modal_error_html_content'));
+                setStepToRender(STEPS.ERROR);
             } else if (!result.filter) {
                 setStepToRender(STEPS.ERROR);
             } else {